Add SequenceMovingWindowReduceTransform
